number of questions now have float type

+++ modules/core/src/main/java/org/vstu/compprehension/dto/ExerciseCardDto.java
@@ -13,7 +13,7 @@ public class ExerciseCardDto {
     @NotNull String strategyId;
     @NotNull String backendId;
     float complexity;
-    int numberOfQuestions;
+    float numberOfQuestions;
     //@NotNull List<String> tags;
     /*
     @NotNull ExerciseOptionsEntity options;
